requirements.txt
fastapi==0.110.0  # Added FastAPI for API backend
uvicorn==0.29.0  # Added Uvicorn to run FastAPI
matplotlib==3.10.0
numpy==1.26.4
opencv-python==4.11.0.86
pandas==2.2.3
Pillow==11.1.0
plotly==6.0.0
PyYAML==6.0.2
requests==2.32.3
tensorflow==2.18.0
torch==2.6.0
torchvision==0.21.0
transformers==4.49.0
ultralytics==8.3.76
